fix: address Codex review comments — スキル名の二重管理をやめる
Codex 指摘（MUST-FIX）: Settings の「Configure appearance…」は grid view と single view の
両方が自前の handler で答えており、slug がそれぞれに直書きされていた。今回の改名で
src/App.vue だけが mulmoterminal-dirs になり、GridView.vue は mulmoterminal-config のまま
残っていたため、同じボタンが「どちらの画面から押したか」で別のスキルを起動していた。
どちらの経路も単体では筋が通っているので、型検査もテストも CI も通っていた。

修正は名前を 1 か所に寄せること。CLAUDE.md が「両側が判断に使う値は common/ に置き、
server/ と src/ にミラーしない」と定めているとおり common/bundledSkills.ts を作り、
BUNDLED_SKILL_NAMES と DIR_CONFIG_SKILL をそこへ移した。install-bundled-skills.ts の
SCHEMA_OWNER_SKILL も同じ定数に統合（どちらも「.mulmoterminal.json を書くスキル」で、
概念としてひとつ）。

Claude 自身のレビューで見つけた同種の取りこぼし（Codex 未指摘）— 同じ「改名がコピーの
片方にしか届かない」型:
- SettingsModal.vue の Keyboard shortcuts 節が /mulmoterminal-config を案内していた
  → keymap を持つのは -keys
- ModelSetupHelp.vue が provider 設定を mulmoterminal-config と案内していた → -model
- DirConfigPreview.vue のコメント → -dirs
- server/index.ts / install-bundled-skills.ts の冒頭コメントを分割後の構成に更新

テスト: DIR_CONFIG_SKILL が実在する同梱スキルを指すこと、その seed が
/mulmoterminal-dirs になること。BUNDLED_SKILL_NAMES の import 元も common/ に寄せた。
